feat(account): pr 36 review comments statistics
